+++ src/nautilus/syscalls/stat.c
@@ -1,12 +1,12 @@
 #include <nautilus/nautilus.h>
 #include <nautilus/fs.h>
 
-int
-sys_stat(int a,int b, int c, int d, int e, int f)
+uint64_t
+sys_stat(uint64_t a,uint64_t b, uint64_t c, uint64_t d, uint64_t e, uint64_t f)
 {
 	// TODO: add protection against bad fd
 	char* pathname = (char*)a;
 	struct nk_fs_stat *st = (struct nk_fs_stat *)b;
-	int ret = nk_fs_stat(pathname,st);
+	uint64_t ret = nk_fs_stat(pathname,st);
 	return ret;
 }
